+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version api packages
 - Temples icons
 
+#### Fixed
+- Display amount on unfinished flow list
+
 ## [0.1.0-rc.3] - 2021-10-19
 #### Added
 - Auctions tab shell